requirements.txt
fastapi==0.95.1          # Ensure you have the appropriate FastAPI version
uvicorn==0.22.0          # For running FastAPI applications

# Google Cloud and BigQuery dependencies
google-cloud-bigquery==3.8.0  # BigQuery library to interact with Google BigQuery
google-cloud-core             # Core library needed for Google services
google-auth==2.20.0           # Authentication library for Google Cloud services

# Pydantic for data validation
pydantic==1.10.2

# Async libraries
httpx==0.24.1           # For async HTTP requests, if needed

# Other optional dependencies, depending on your project
sqlalchemy==2.0.16      # If you are using SQLAlchemy for ORM
python-dotenv==1.0.0    # For environment variables

# Cloud logging (Optional)
google-cloud-logging==3.2.0
